+++ app/index.js
@@ -1,5 +1,18 @@
 import { be_sig_triple_ } from 'ctx-core/rmemo'
-import { app_ctx__be_config } from 'rebuildjs'
+import { Elysia } from 'elysia'
+import { staticPlugin } from '@elysiajs/static'
+import { join } from 'path'
+import {
+	app_ctx,
+	app_ctx__be_config,
+	browser_path_,
+	cwd_,
+	middleware_ctx_,
+	server__input_path__set,
+	server__metafile_,
+	server__output_,
+	server__output_path_
+} from 'rebuildjs'
 export * from 'rebuildjs/app'
 export const [
 	app$_,
@@ -8,3 +21,26 @@ export const [
 ] = be_sig_triple_(()=>
 	undefined,
 app_ctx__be_config)
+export async function app__new() {
+	const server__metafile = server__metafile_(app_ctx)
+	const app =
+		new Elysia().use(staticPlugin({
+			assets: browser_path_(app_ctx),
+			prefix: '',
+		}))
+	if (server__metafile) {
+		const { inputs } = server__metafile
+		for (let input_path in inputs) {
+			const middleware_ctx = middleware_ctx_()
+			server__input_path__set(middleware_ctx, input_path)
+			const output = server__output_(middleware_ctx)
+			if (output) {
+				const server__middleware =
+					await import(join(cwd_(app_ctx), server__output_path_(middleware_ctx)))
+						.then(mod=>mod.default)
+				app.use(server__middleware(middleware_ctx))
+			}
+		}
+	}
+	return app
+}